requirements.txt
dnspython>=1.14.0
requests>=2.20.0
ppdeep>=20200505
tld>=0.9.1
certifi>=2020.06.20
scikit-image>=0.17.0
netaddr
lxml
selenium
webdriver-manager
matplotlib
opencv-python-headless
numpy
python-nmap
argparse
dnstwist==20201228
progress
reconDNS>=0.0.3
whoisdomain==1.20230720.2
fake_useragent
ultralytics
pillow
